refactor: split sync.py into modular package structure
Split the 3069-line sync.py into a well-organized package:
- sync/__init__.py - Router aggregation
- sync/dbt.py - dbt manifest sync (~1700 lines)
- sync/openapi.py - OpenAPI import/impact/diff (~500 lines)
- sync/graphql.py - GraphQL import/impact/diff (~500 lines)
- sync/helpers.py - Shared utilities (resolve_team_by_name, resolve_user_by_email)

All 840 tests pass.
